四叉树链路尺寸层 64 位化，修复超大边长上限整数除零
- 坐标类型：新增 Point2l / Rect2l 别名与 point_to_l（沿用 std::llround），
  point_to_double 补 Point2l 重载；Point2i / Rect2i / point_to_int 语义不变
- 分层约定：对外公开接口（Quadtree_Manager::seek 单点与范围、qurdtree_build_smart、
  set_max_size 等）保持 Point2i / Rect2i；模块内部的树边界范围、由尺寸推导的坐标、
  待查询区块坐标、树间范围计算全部换 Point2l / Rect2l；callback 契约的 target
  参数是申请扩大的目标坐标，随之升级为 Point2l
- 四叉树：函数预声明.h 与 core/导航与范围计算.hpp、core/区块信息检索.hpp、
  core/设置与交互.hpp 同步换型；manage_range_calcu 全程 int64 运算后落 int64 分量；
  child_node_range_calcu 中点改 int64；target_range_format 补 block_size == 0 早退
- 四叉树管理器：函数预声明.h 与六个 core 实现同步换型（四叉树智能创建、区块信息检索、
  基础操作、相邻四叉树查找、四叉树扩大回调管理、四叉树合并）；
  tree_cache.ranges 改 std::vector<Rect2l>；尺寸与区块索引系列一律 int64_t
- 主崩点修复：prepare_smart_create_params 原以 int 承载 settings.max_tree_size，
  边长上限 2^33 按模 2^32 截断为 0，紧随其后的 width % father_block_size
  抛 SEH 0xC0000094；现改用 int64_t 并补 father_block_size <= 0 早退。
  该机制已用最小复现程序实测确认：2^33 存入 int 得 0，1 % 0 退出码 0xC0000094
- 同类隐患：excel_element_to_coord 补 block_size == 0 与 width <= 0 防御；
  range_seek 的 total_num 乘法前做溢出检查，越界直接返回；建档前拦截超出 int 的坐标
- 已知限制：超大树上的范围查询仍受内存物理限制，本版次只承诺建树与单点检索

引擎层 Debug 构建：configure exit=0、build exit=0、EngineCore.lib 产出，
error C / error LNK 均为 0，告警与既有基线一致。

+++ src/core/spatial/partition/Quadtree/core/设置与交互.hpp
@@ -97,7 +97,7 @@ namespace engine
 	//四叉树回调管理方法设置
 	template <typename T>
 	void Quadtree<T>::set_callback_manage
-	(const std::function<bool(Point2d root, Point2i target)>& cb)
+	(const std::function<bool(Point2d root, Point2l target)>& cb)
 	{
 		//回调管理函数注册
 		callback = cb;
