+++ src/core/spatial/partition/Quadtree/函数预声明.h
@@ -59,14 +59,17 @@ namespace engine
 			//节点
 			Node* node;
 			//节点范围
-			Rect2i node_range{};
+			//范围边界由树边长推导，边长可配置到 INT_MAX 以上
+			//故一律使用 64 位整数承载，避免溢出
+			Rect2l node_range{};
 			//递归级别
 			int recur_level = 0;
 		};
 		//四叉树状态记录
 		tree_state state;
 		//外界上级管理对象回调管理方法----四叉树扩大行为权限申请
-		std::function<bool(const Point2d& root, const Point2i& target)> callback;
+		//目标坐标取自范围推导结果，同样按 64 位整数通报
+		std::function<bool(const Point2d& root, const Point2l& target)> callback;
 
 		// ———— 公开接口 ————
 	public:
@@ -82,13 +85,14 @@ namespace engine
 		void set_max_size(const uint64_t& size);
 		//四叉树回调管理方法设置
 		void set_callback_manage
-		(const std::function<bool(Point2d root, Point2i target)>& cb);
+		(const std::function<bool(Point2d root, Point2l target)>& cb);
 
 		// ---- 查询 ----
 		//最小区块单元查找
-		void block_seek(tree_chunk_data<T>*& receiver, const Point2i& target, bool stable);
+		//目标坐标为待查询区块坐标，来自范围推导，按 64 位整数接收
+		void block_seek(tree_chunk_data<T>*& receiver, const Point2l& target, bool stable);
 		//范围区块单元查找
-		void range_seek(std::vector<tree_chunk_data<T>*>& receiver, const Rect2i& target_range, bool stable);
+		void range_seek(std::vector<tree_chunk_data<T>*>& receiver, const Rect2l& target_range, bool stable);
 
 		//四叉树状态获取
 		const tree_state& tree_state_get(void);
@@ -102,23 +106,23 @@ namespace engine
 		void recur_level_calcu(int& address_series);
 
 		//递归方向计算
-		void recur_direct_calcu(const Point2i& target, const Rect2i& node, int& recur_direct);
+		void recur_direct_calcu(const Point2l& target, const Rect2l& node, int& recur_direct);
 
 		//子节点范围计算
-		void child_node_range_calcu(const int& recur_direct, Rect2i& child_range,
-			const Rect2i& parent_range);
+		void child_node_range_calcu(const int& recur_direct, Rect2l& child_range,
+			const Rect2l& parent_range);
 	public:
 		//四叉树管理范围计算
-		void manage_range_calcu(Rect2i& receiver,const Point2d& root,const uint64_t tree_size);
+		void manage_range_calcu(Rect2l& receiver,const Point2d& root,const uint64_t tree_size);
 
 		//待查询范围格式化
-		void target_range_format(Rect2i& target_range,const Point2d& root, const uint64_t block_size);
+		void target_range_format(Rect2l& target_range,const Point2d& root, const uint64_t block_size);
 
 		//可查询范围计算
-		Point2d seekable_range_calcu(const Rect2i& target_range, Rect2i& seekable_range);
+		Point2d seekable_range_calcu(const Rect2l& target_range, Rect2l& seekable_range);
 	private:
 		//查询范围关系获取
-		bool range_relation_get(const Rect2i& target_range, const Rect2i& node_range);
+		bool range_relation_get(const Rect2l& target_range, const Rect2l& node_range);
 
 		// ———— 结构维护 ————
 	private:
@@ -131,14 +135,14 @@ namespace engine
 		// ———— 查询前置支撑 ————
 	private:
 		//单点查询可行性分析
-		int point_seekable_analyse(const Point2i& target);
+		int point_seekable_analyse(const Point2l& target);
 
 		//范围查询可行性分析
-		void range_seekable_analyse(const Rect2i& format_range, Rect2i& seekable_range);
+		void range_seekable_analyse(const Rect2l& format_range, Rect2l& seekable_range);
 
 		//递归栈操作
 		void recur_stack_operate(std::vector<recur_record>& recur_stack,
-			Node*& ptr, Rect2i& range, int& level,
+			Node*& ptr, Rect2l& range, int& level,
 			bool push_back);
 
 	};
