+++ src/core/spatial/partition/Quadtree_Manager/core/基础操作.hpp
@@ -6,10 +6,10 @@ namespace engine
 {
     //直属四叉树查找
     template<typename T>
-    tree_record<T>* Quadtree_Manager<T>::quadtree_inclusion_seek(const Point2i& target)
+    tree_record<T>* Quadtree_Manager<T>::quadtree_inclusion_seek(const Point2l& target)
     {
         //临时树边界存储
-        Rect2i tree_range{};
+        Rect2l tree_range{};
 
         //简化表示路径
         auto& is_cache_enabled = settings.is_cache_enabled;
@@ -154,7 +154,8 @@ namespace engine
         //设置新四叉树最小区块单元大小
         new_tree->tree->set_block_size(settings.block_size);
         //设置新四叉树回调管理函数
-        auto manage = [this](const Point2d& root, const Point2i& seek)->bool
+        //申请坐标按 64 位整数传递，与四叉树内部坐标系一致
+        auto manage = [this](const Point2d& root, const Point2l& seek)->bool
             { return this->tree_expand_approve(root, seek); };
         new_tree->tree->set_callback_manage(manage);
